Move container's send_action to http transport
