+++ skygear/bin.py
@@ -86,7 +86,7 @@ def run_plugin(options):
             "Connecting to address %s" % options.skygear_address)
         transport = ZmqTransport(options.skygear_address,
                                  threading=options.zmq_thread_pool)
-        SkygearContainer.set_bidirectional_transport(transport)
+    SkygearContainer.set_default_transport(transport)
     transport.run()
